+++ client/src/app/ListView.tsx
@@ -48,6 +48,9 @@ const ListView = ({ locations }: { locations: Restaurant[] }) => {
           <div className="text-xs text-gray-300">
             Listing updated at {new Date(location.updatedAt).toDateString()}
           </div>
+          <div>
+            <b>Status: </b> {location.status}
+          </div>
           <p>
             <b>Address:</b> {location.address}
           </p>