+++ client/src/app/Map.tsx
@@ -41,7 +41,7 @@ const Map = ({ locations }: { locations: Restaurant[] }) => (
               Listing updated at {new Date(location.updatedAt).toDateString()}
             </div>
             {location.status !== "Operational" && (
-              <div>
+              <div className="bg-red-800 text-white">
                 <b>Status: </b> {location.status}
               </div>
             )}