+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Display restaurant status in UI (either Operational, Temporarily Closed, Permanently Closed)
+
 ### Updated
 
 - Add more padding around items in list view